LEARNIX-69 In LectionProgress den progress auf Integer umgestellt

+++ src/main/java/de/szut/learnixback/dto/LectureProgressDTO.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 public class LectureProgressDTO {
     private String userGUID;
     private Long lectionID;
-    private Float progress;
+    private Integer progress;
 }
